+++ zapisy/apps/help/views.py
@@ -4,45 +4,53 @@
     Help views
 """
 
-from django.template import RequestContext
-from django.http import HttpResponse, HttpResponseRedirect
-from django.shortcuts import render_to_response
+from django.shortcuts import render
 
 from django.contrib.admin.views.decorators import staff_member_required
 from apps.users.decorators import employee_required
 
-def main_page( request ):
-    return render_to_response('help/base.html', {}, context_instance = RequestContext(request))
 
-def terms( request ):
-    return render_to_response('help/terms.html', {}, context_instance = RequestContext(request))
+def main_page(request):
+    return render(request, 'help/base.html', {})
 
-def rules( request ):
-    return render_to_response('help/rules.html', {}, context_instance = RequestContext(request))
 
-def enrollment( request ):
-    return render_to_response('help/enrollment.html', {}, context_instance = RequestContext(request))
+def terms(request):
+    return render(request, 'help/terms.html', {})
 
-def grade( request ):
-    return render_to_response('help/grade.html', {}, context_instance = RequestContext(request))
 
-def mobile( request ):
-    return render_to_response('help/mobile.html', {}, context_instance = RequestContext(request))
+def rules(request):
+    return render(request, 'help/rules.html', {})
 
-def export( request ):
-    return render_to_response('help/export.html', {}, context_instance = RequestContext(request))
 
-def offer( request ):
-    return render_to_response('help/offer.html', {}, context_instance = RequestContext(request))
+def enrollment(request):
+    return render(request, 'help/enrollment.html', {})
 
-def errorpage( request ):
+
+def grade(request):
+    return render(request, 'help/grade.html', {})
+
+
+def mobile(request):
+    return render(request, 'help/mobile.html', {})
+
+
+def export(request):
+    return render(request, 'help/export.html', {})
+
+
+def offer(request):
+    return render(request, 'help/offer.html', {})
+
+
+def errorpage(request):
     raise Exception('Example exception')
 
+
 @staff_member_required
-def admin( request ):
-    return render_to_response('help/admin.html', {}, context_instance = RequestContext(request))
+def admin(request):
+    return render(request, 'help/admin.html', {})
+
 
 @employee_required
-def employee( request ):
-    return render_to_response('help/employee.html', {}, context_instance = RequestContext(request))
-
+def employee(request):
+    return render(request, 'help/employee.html', {})