+++ zapisy/apps/offer/desiderata/views.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 from django.http import Http404
-from django.shortcuts import render_to_response
-from django.template import RequestContext
+from django.shortcuts import render
 from django.contrib import messages
 
 from apps.offer.desiderata.models import Desiderata, DesiderataOther
@@ -42,7 +41,7 @@ def change_desiderata(request):
         'other_form': other_form,
         'semester': semester
     }
-    return render_to_response('offer/desiderata/change_desiderata.html', data, context_instance=RequestContext(request))
+    return render(request, 'offer/desiderata/change_desiderata.html', data)