+++ zapisy/apps/offer/preferences/views.py
@@ -1,25 +1,21 @@
 # -*- coding:utf-8 -*-
+import logging
 
 from django.core.exceptions import ObjectDoesNotExist
+from django.shortcuts import render
+from django.views.decorators.http import require_POST
 
-from django.shortcuts import render_to_response
-from django.template import RequestContext
-
+from libs.ajax_messages import AjaxFailureMessage
 from apps.offer.preferences.forms import PreferenceFormset, PreferenceForm
 from apps.offer.preferences.models import Preference
 from apps.users.decorators import employee_required
 from apps.enrollment.courses.models import Semester
 
-from django.views.decorators.http import require_POST
-
-import logging
-from libs.ajax_messages import AjaxSuccessMessage, AjaxFailureMessage
-
 logger = logging.getLogger()
+
 
 @employee_required
 def view(request):
-
     employee = request.user.employee
     employee.make_preferences()
 
@@ -27,10 +23,7 @@ def view(request):
     formset   = PreferenceFormset(queryset=prefs)
     semester  = Semester.get_current_semester()
 
-    return render_to_response(
-        'offer/preferences/base.html',
-        locals(),
-        context_instance = RequestContext(request))
+    return render(request, 'offer/preferences/base.html', locals())
 
 
 @employee_required
@@ -60,5 +53,5 @@ def save(request):
             setattr(pref, field, form[field].value())
         pref.save()
         form = PreferenceForm(instance=pref)
-        return render_to_response('offer/preferences/form_row.html', {'form': form, })
+        return render(request, 'offer/preferences/form_row.html', {'form': form, })
     return AjaxFailureMessage('InvalidRequest', u'Coś poszło źle')