+++ zapisy/apps/users/views.py
@@ -9,26 +9,25 @@ from django.contrib.auth.models import User
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.decorators import permission_required
 from django.contrib.auth.views import login
-from django.shortcuts import render_to_response, redirect
-from django.template import RequestContext
+from django.shortcuts import render, redirect
 from django.core.urlresolvers import reverse
 from django.http import HttpResponse, HttpResponseRedirect
 from django.template.response import TemplateResponse
 from django.utils.translation import check_for_language, LANGUAGE_SESSION_KEY
 from django.conf import settings
 
-import vobject
+from vobject import iCalendar
 
+from apps.enrollment.records.utils import prepare_schedule_courses, prepare_schedule_data
 from apps.grade.ticket_create.models.student_graded import StudentGraded
 from apps.offer.vote.models.single_vote import SingleVote
 from apps.enrollment.courses.exceptions import MoreThanOneCurrentSemesterException
 from apps.users.utils import prepare_ajax_students_list, prepare_ajax_employee_list
-from apps.users.models import Employee, Student, BaseUser
+from apps.users.models import Employee, Student, BaseUser, UserProfile, OpeningTimesView
 from apps.enrollment.courses.models import Semester, Group
 from apps.enrollment.records.models import Record
 from apps.enrollment.utils import mailto
 from apps.users.forms import EmailChangeForm, BankAccountChangeForm, ConsultationsChangeForm, EmailToAllStudentsForm
-from apps.enrollment.records.utils import *
 from apps.notifications.forms import NotificationFormset
 from apps.notifications.models import NotificationPreferences
 from libs.ajax_messages import AjaxSuccessMessage
@@ -40,6 +39,7 @@ GTC = {'1': 'wy', '2': 'cw', '3': 'pr',
        '4': 'cw', '5': 'cw+prac',
        '6': 'sem', '7': 'lek', '8': 'WF',
        '9': 'rep', '10': 'proj'}
+BREAK_DURATION = datetime.timedelta(minutes=15)
 
 
 @login_required
@@ -65,13 +65,13 @@ def student_profile(request, user_id):
         })
 
         if request.is_ajax():
-            return render_to_response('users/student_profile_contents.html', data, context_instance=RequestContext(request))
+            return render(request, 'users/student_profile_contents.html', data)
         else:
             students = Student.get_list()
             enrolled_students = Record.recorded_students(students)
             data['students'] = enrolled_students
             data['char'] = "All"
-            return render_to_response('users/student_profile.html', data, context_instance=RequestContext(request))
+            return render(request, 'users/student_profile.html', data)
 
     except Student.DoesNotExist:
         logger.error(
@@ -79,14 +79,14 @@ def student_profile(request, user_id):
             .format(str(user_id))
         )
         messages.error(request, "Nie ma takiego studenta.")
-        return render_to_response('common/error.html', context_instance=RequestContext(request))
+        return render(request, 'common/error.html')
     except User.DoesNotExist:
         logger.error(
             'Function student_profile(id = {}) throws User.DoesNotExist while acessing to non existing user.'
             .format(str(user_id))
         )
         messages.error(request, "Nie ma takiego użytkownika.")
-        return render_to_response('common/error.html', context_instance=RequestContext(request))
+        return render(request, 'common/error.html')
 
 def employee_profile(request, user_id):
     """employee profile"""
@@ -100,11 +100,11 @@ def employee_profile(request, user_id):
     except Employee.DoesNotExist:
         logger.error('Function employee_profile(user_id = %s) throws NonEmployeeException while acessing to non existing employee.' % str(user_id) )
         messages.error(request, "Nie ma takiego pracownika.")
-        return render_to_response('common/error.html', context_instance=RequestContext(request))
+        return render(request, 'common/error.html')
     except User.DoesNotExist:
         logger.error('Function employee_profile(id = %s) throws User.DoesNotExist while acessing to non existing user.' % str(user_id) )
         messages.error(request, "Nie ma takiego użytkownika.")
-        return render_to_response('common/error.html', context_instance=RequestContext(request))
+        return render(request, 'common/error.html')
 
     try:
         courses = prepare_schedule_courses(request, for_employee=employee)
@@ -115,8 +115,7 @@ def employee_profile(request, user_id):
         })
 
         if request.is_ajax():
-            return render_to_response('users/employee_profile_contents.html',
-                data, context_instance=RequestContext(request))
+            return render(request, 'users/employee_profile_contents.html', data)
         else:
             semester = Semester.get_current_semester()
             employees = Employee.get_list()
@@ -129,14 +128,14 @@ def employee_profile(request, user_id):
             data['employees'] = active_employees
             data['char'] = 'All'
 
-            return render_to_response('users/employee_profile.html', data, context_instance=RequestContext(request))
+            return render(request, 'users/employee_profile.html', data)
 
 
     except MoreThanOneCurrentSemesterException:
         data = {'employee' : employee}
         logger.error('Function employee_profile throws MoreThanOneCurrentSemesterException.' )
         messages.error(request, "Przepraszamy, system jest obecnie nieaktywny z powodu niewłaściwej konfiguracji semestrów. Prosimy spróbować później.")
-        return render_to_response('users/employee_profile.html', data, context_instance=RequestContext(request))
+        return render(request, 'users/employee_profile.html', data)
 
 @login_required
 def set_language(request):
@@ -178,7 +177,7 @@ def email_change(request):
 
             if user and user <> request.user:
                 messages.error(request, "Podany adres jest już przypisany do innego użytkownika!")
-                return render_to_response('users/email_change_form.html', {'form':form}, context_instance=RequestContext(request))
+                return render(request, 'users/email_change_form.html', {'form':form})
 
             form.save()
             logger.info('User (%s) changed email' % request.user.get_full_name())
@@ -186,7 +185,7 @@ def email_change(request):
             return HttpResponseRedirect(reverse('my-profile'))
     else:
         form = EmailChangeForm({'email' : request.user.email})
-    return render_to_response('users/email_change_form.html', {'form':form}, context_instance=RequestContext(request))
+    return render(request, 'users/email_change_form.html', {'form':form})
 
 @login_required
 def bank_account_change(request):
@@ -203,7 +202,7 @@ def bank_account_change(request):
     else:
         zamawiany = Student.get_zamawiany(request.user.id)
         form = BankAccountChangeForm({'bank_account': zamawiany.bank_account})
-    return render_to_response('users/bank_account_change_form.html', {'form':form}, context_instance=RequestContext(request))
+    return render(request, 'users/bank_account_change_form.html', {'form':form})
 
 @login_required
 def consultations_change(request):
@@ -220,11 +219,10 @@ def consultations_change(request):
                 return HttpResponseRedirect(reverse('my-profile'))
         else:
             form = ConsultationsChangeForm({'consultations': employee.consultations, 'homepage': employee.homepage, 'room': employee.room})
-        return render_to_response('users/consultations_change_form.html', {'form':form}, context_instance=RequestContext(request))
+        return render(request, 'users/consultations_change_form.html', {'form':form})
     except Employee.DoesNotExist:
         messages.error(request, 'Nie jesteś pracownikiem.')
-        return render_to_response('common/error.html',
-                context_instance=RequestContext(request))
+        return render(request, 'common/error.html')
 
 @login_required
 def password_change_done(request):
@@ -295,7 +293,7 @@ def employees_list(request, begin = 'All', query=None):
             "query": query
             }
 
-    return render_to_response('users/employees_list.html', data, context_instance=RequestContext(request))
+    return render(request, 'users/employees_list.html', data)
 
 def consultations_list(request, begin='A'):
 
@@ -311,7 +309,7 @@ def consultations_list(request, begin='A'):
             "employees" : employees,
             "char": begin
             }
-        return render_to_response('users/consultations_list.html', data, context_instance=RequestContext(request))
+        return render(request, 'users/consultations_list.html', data)
 
 
 @login_required
@@ -330,7 +328,7 @@ def students_list(request, begin = 'All', query=None):
             'mailto_group': mailto(request.user, students),
             'mailto_group_bcc': mailto(request.user, students, True)
         }
-        return render_to_response('users/students_list.html', data, context_instance=RequestContext(request))
+        return render(request, 'users/students_list.html', data)
 
 @login_required
 def logout(request):
@@ -367,7 +365,7 @@ def create_ical_file(request):
     user_full_name = user.get_full_name()
     semester = Semester.get_default_semester()
 
-    cal = vobject.iCalendar()
+    cal = iCalendar()
     cal.add('x-wr-timezone').value = 'Europe/Warsaw'
     cal.add('version').value = '2.0'
     cal.add('prodid').value = 'Fereol'
@@ -395,6 +393,7 @@ def create_ical_file(request):
             continue
         for term in terms:
             start_datetime = datetime.datetime.combine(term.day, term.start)
+            start_datetime += BREAK_DURATION
             end_datetime = datetime.datetime.combine(term.day, term.end)
             event = cal.add('vevent')
             event.add('summary').value = '%s - %s' % (course_name, group_type)
@@ -442,4 +441,4 @@ def email_students(request):
     else:
         form = EmailToAllStudentsForm(initial={'sender': 'zapisy@cs.uni.wroc.pl'})
         form.fields['sender'].widget.attrs['readonly'] = True
-    return render_to_response('users/email_students.html', {'form':form, 'students_mails': studentsmails}, context_instance=RequestContext(request))
+    return render(request, 'users/email_students.html', {'form':form, 'students_mails': studentsmails})