+++ zapisy/pipeline/storage.py
@@ -32,7 +32,7 @@ class PipelineStorage(FileSystemStorage):
     def modified_time(self, name):
         return datetime.fromtimestamp(os.path.getmtime(self.path(name)))
 
-    def get_available_name(self, name):
+    def get_available_name(self, name, max_length = None):
         if self.exists(name):
             self.delete(name)
         return name